packages/tele-bot/pyproject.toml
[project]
name = "tele-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "ZSHenChan",email = "zishenchan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai (>=1.60.0,<2.0.0)",
    "telethon (>=1.42.0,<2.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pyobjc-framework-eventkit (>=12.1,<13.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
